refactor: piece 패키징
