+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.piece;
 
+import chess.domain.Move;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
